+++ src/components/index.js
@@ -1,4 +1,7 @@
 export { Btn } from './Btn/Btn';
+export { ProtectedRoute } from './ProtectedRoute/ProtectedRoute';
+export { AdviseBox } from './AdviseBox/AdviseBox';
+export { AppRoutes } from './AppRoutes/AppRoutes';
 
 // Header
 export { UserInfo } from './UserInfo/UserInfo';
