+++ src/pages/index.js
@@ -1,0 +1,6 @@
+export { AllTracs } from 'pages/AllTracs/AllTracs';
+export { EntryForm } from 'pages/EntryForm/EntryForm';
+export { Favorites } from 'pages/Favorites/Favorites';
+export { MainPage } from 'pages/MainPage/MainPage';
+export { NotFound } from 'pages/NotFound/NotFound';
+export { Playlist } from 'pages/Playlist/Playlist';
